Move WeightInfo, Config, Event into relayer pallet.
