+++ pallets/runtime/tests/src/staking/mock.rs
@@ -43,13 +43,15 @@ use pallet_staking::types::SlashingSwitch;
 use pallet_staking::{self as pallet_staking, *};
 use polymesh_common_utilities::traits::balances::{AccountData, CheckCdd};
 use polymesh_common_utilities::traits::portfolio::PortfolioSubTrait;
-use polymesh_common_utilities::traits::relayer::SubsidiserTrait;
 use polymesh_common_utilities::traits::CommonConfig;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::constants::currency::POLY;
 use polymesh_primitives::identity_id::GenesisIdentityRecord;
 use polymesh_primitives::{
-    traits::group::{GroupTrait, InactiveMember},
+    traits::{
+        group::{GroupTrait, InactiveMember},
+        SubsidiserTrait,
+    },
     Authorization, AuthorizationData, Claim, IdentityId, Moment, NFTId, Permissions, PortfolioId,
     SecondaryKey, Signatory,
 };
